pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "heimdallr"
version = "1.0.0"
description = "AI 诊断助手 - 基于 CrewAI 的智能诊断系统，支持动态工作流和多类型输入"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
authors = [
    {name = "ruhuo.tan", email = "ruhuo.tan@shopee.com"}
]
keywords = ["fastapi", "ai", "diagnosis", "crewai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "crewai>=0.150.0",
    "crewai-tools>=0.58.0",
    "python-dotenv>=1.1.1",
    "langchain-openai>=0.2.14",
    "langchain-google-genai>=2.1.8",
    "fastapi>=0.116.1",
    "uvicorn>=0.35.0",
    "httpx>=0.28.1",
    "jira>=3.10.5",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

[project.urls]
Homepage = "https://git.garena.com/shopee/loan-service/credit_frontend/cashloan/heimdallr"
Repository = "https://git.garena.com/shopee/loan-service/credit_frontend/cashloan/heimdallr"
Issues = "https://git.garena.com/shopee/loan-service/credit_frontend/cashloan/heimdallr/issues"

[project.scripts]
heimdallr = "main:app"

[tool.hatch.build.targets.wheel]
packages = ["app"]
include = [
    "main.py",
    "app/**/*.py",
]

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "crewai.*",
    "crewai_tools.*",
    "jira.*",
]
ignore_missing_imports = true 